refactor: unify viewModel
