+++ feature/feature-record/src/main/java/com/example/feature_record/DailyGoodsSalesSummary.kt
@@ -1,0 +1,6 @@
+package com.example.feature_record
+
+data class DailyGoodsSalesSummary(
+    val date: String,
+    val salesSummary: List<GoodsSalesSummary>
+)
