+++ feature/feature-record/src/main/java/com/example/feature_record/GoodsSalesSummary.kt
@@ -1,0 +1,7 @@
+package com.example.feature_record
+
+data class GoodsSalesSummary(
+    val goodsId: Long,
+    val goodsName: String,
+    var totalQuantity: Int = 0
+)
